move io and web foundation types into libraries
